+++ bin/rename.dart
@@ -1,4 +1,4 @@
-import 'package:xly/src/app_renamer.dart';
+import '../tool/app_renamer.dart';
 
 void main(List<String> args) async {
   await AppRenamer.main(args);